+++ Assets/1.Scripts/Game/ObjectPool/HitPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4155428d5ee08f741a10ed363e1a5abb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
