+++ Assets/1.Scripts/Game/Weapon/Weapon.cs
@@ -145,15 +145,3 @@ public class Weapon : MonoBehaviour
         _shells.Enqueue(cartridge);
     }
 }
-
-public class HitPoint
-{
-    public Vector3 HitPosition = Vector3.zero;
-    public Vector3 HitNormal = Vector3.zero;
-
-    public void Init(RaycastHit hit)
-    {
-        HitPosition = hit.point;
-        HitNormal = hit.normal;
-    }
-}
